requirements.txt
streamlit==1.28.0
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
plotly==5.17.0  # برای مصورسازی
pandas==2.0.3   # برای مدیریت داده